WORKAROUND: Add __dso_handle declaration
This needs to be investigated further... the workaround is probably safe
since we are not using shared objects.

Something has probably changed in the Makefiles since 2.6.29 because this
symbol was getting pulled in from the GCC libraries there.  Not sure why
this is not happening any more... also not quite sure which is the correct
behaviour.