Make xdrproc_t non-variadic
Casting a non-variadic function to a variadic type is not permitted in C and
results in undefined behaviour.  Although it may work for some architectures,
on the OR32 architecture (and for others, from what I can find through Google),
the parameters are passed differently depending on whether the function is
declared varidic or not, thus casting between the two types results in
failure.

Specifically, the following is not permitted:

/* Define non-variadic function */
int foo(int x, int y) {
	...
}

typedef int (*bar_t)(int x, int y, ...);

void f() {
	int x,y;
	bar_t bar;

	/* Cast foo to variadic type... not allowed */
	bar = (bar_t) foo;
	(*bar)(x,y);
}

This patch redeclares the xdrproc_t to be non-variadic.  None of the functions
that are cast to xdrproc_t in uClibc are actually variadic; this applies to
Busybox, as well.