Enable global constructors/destructors
C++ requires global constructors in order to run the necessary _init
functions.

As this is urgent and it's not clear how to get crt1.S to build, the
init and fini functions are being passed to uClibc_main in crt0.S.
Hopefully this doesn't have side effects and is good enough for now
until this can be revisited and done "properly".

Patch provided by Jeremy Bennett.