Adapt to new Linux syscall ABI
The Linux syscall ABI has finally settled in OpenRISC Linux 2.6.38.  This
new ABI is, externally, the same as the regular C function call ABI with
a couple of exceptions:

i)  r9 does not need to be set as the link register is not used for
    returning from the user interrupt; r9 is still preserved across call

ii) r11 is used when entering the user interrupt to determine which
    syscall to execute; r11 will contain function's return value upon
    return (i.e. the syscall result)

Aside from those exceptions, a syscall should behave like a regular function
call with all that implies with respect to call-saved and call-clobbered
registers.

For completeness, the register characteristics are:

call clobbered: r3-r8, r11-r13, and odd numbered regs from r15-r31
call saved:     r1,r2,r9,r10, and even numbered regs from r14-r30